// webui/src/core/ConfigurationProvider/types.ts
export type ConfigurationId = Branded<string, "ConfigurationId">;

export function createNewConfigurationId() {
	return v4() as ConfigurationId;
}

export type AttributeConfig = {
	/**
	 * Name of the attribute in the DB schema
	 */
	name: string;

	/**
	 * DB data type
	 */
	dataType?: string;
};

export type VertexTypeConfig = {
	/**
	 * Vertex type name's in the DB schema
	 */
	type: VertexType;
	/**
	 * Name to be printed
	 */
	displayLabel?: string;
	/**
	 * Optional icon to be rendered inside the graph viewer and
	 * other modules.
	 */
	iconUrl?: string;
	/**
	 * Vertex attribute to be used as label
	 */
	displayNameAttribute?: string;
	/**
	 * Vertex attribute to be used as description
	 */
	longDisplayNameAttribute?: string;
	/**
	 * List of attributes for the vertex type
	 */
	attributes: Array<AttributeConfig>;
	/**
	 * Total number of vertices of this type
	 */
	total?: number;
} & VertexStyleStorage;

export type EdgeTypeConfig = {
	/**
	 * Edge type name's in the DB schema
	 */
	type: EdgeType;
	/**
	 * Name to be printed
	 */
	displayLabel?: string;
	/**
	 * Vertex attribute to be used as label
	 */
	displayNameAttribute?: string;
	/**
	 * List of attributes for the edge type
	 */
	attributes: Array<AttributeConfig>;
	/**
	 * Total number of edges of this type
	 */
	total?: number;
} & EdgeStyleStorage;

/**
 * Represents a connection between node labels via an edge type.
 * Used by Schema Explorer to visualize relationships between node types.
 */
export type EdgeConnection = {
	/**
	 * The edge type name
	 */
	edgeType: EdgeType;
	/**
	 * The source node label
	 */
	sourceVertexType: VertexType;
	/**
	 * The target node label
	 */
	targetVertexType: VertexType;
	/**
	 * Count of edges with this connection pattern
	 */
	count?: number;
};